Add Ability to Get Seal Status
Also adds a wrapper around node-vault to simplify API requests.

At the moment all these things are on the index page. I'm not sure how to arcitect this at the
moment so i'll contunue to let it evolve. Also, vault server info is kept in vaultProvider at the moment. That needs to be set in the source in order to work. This will need to be taken in on a settings page and probably persisted to disk.