Add Authentication Status to NavBar
This is the simplest implementation of authentication status imaginable.
At the moment, this status ONLY indicated whether the vault client thinks it's authenticated,
meaning whether or not the instance of node-vault has a `token` property that is not `undefined`.
This is the MVP, in the future this will get much more advance but is mostly correct right now
because we only allow authentication via the root token.