Add Tree for Reading and Listing Secrets
Adds an intuitive tree for navigating and reading secrets.
It is much like a directory navigator.
Behind the scenes a GET request is made in real time for each
'directory' expansion and the results are addd to the local data store.
Upon selecting a 'leaf' (secret) it's value is displayed in an area to
the right.
Each time a directory is re-opened its children are reloaded.

This feature replaces `SecretLister` as well as `SecretReader`, so
those views were removed.